+++ sakura_core/io/CBinaryStream.h
@@ -9,6 +9,7 @@
 #define SAKURA_CBINARYSTREAM_856F71C1_27E0_4075_BA32_245D18E142BA_H_
 #pragma once
 
+#include "StdAfx.h"
 #include "CStream.h"
 
 class CBinaryInputStream final : public CStream{
@@ -16,10 +17,10 @@ public:
 	explicit CBinaryInputStream(LPCWSTR pszFilePath);
 
 	//! ストリームの「残り」サイズを取得
-	int GetLength();
+	LONGLONG GetLength();
 
 	//! データを無変換で読み込む。戻り値は読み込んだバイト数。
-	int Read(void* pBuffer, size_t nSizeInBytes);
+	LONGLONG Read(void* pBuffer, size_t nSizeInBytes);
 };
 
 class CBinaryOutputStream final : public COutputStream{
